🐛 [-bug] Fixed issue with empty StreamDecorators and capabilities
Default Capabilities are created for empty StreamDecorators (e.g. `StreamDecorator(None)`). However, the default Capabilities constructor looks at environment variables to override values when none are provided. In this case, we don't want to look at those values and use the lowest settings available.

The fix is to ignore the environment when creating the Capabilities to associate with the empty StreamDecorator.